third_party/WebKit/LayoutTests/platform/android/svg/custom/animation-values-parsing-error-expected.txt

CONSOLE ERROR: Error: Invalid value for <animate> attribute values=";;"
CONSOLE ERROR: Error: Invalid value for <animate> attribute values="300;255;180;30;;"
CONSOLE ERROR: Error: Invalid value for <animate> attribute values="50%;;;;10%"
Tests parsing of values attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
